+++ src/components/item-card/card.js
@@ -72,7 +72,7 @@ export const Card = (props) => {
     read_time,
     openExternal,
     syndicated,
-    original_url,
+    open_url,
     noImage
   } = item
 
@@ -126,7 +126,7 @@ export const Card = (props) => {
       ref={articleRef}
       className={card}
       key={id}
-      data-id={`article-card-${id}`}
+      data-cy={`article-card-${id}`}
       onClick={selectBulk}>
       <div className="selectedBack" />
 
@@ -170,7 +170,7 @@ export const Card = (props) => {
               //eslint-disable-next-line
               <a
                 className="publisher"
-                href={original_url}
+                href={open_url}
                 target="_blank"
                 onClick={onOpenOriginalUrl}
                 tabIndex={0}>