+++ src/connectors/global-nav/global-nav.js
@@ -71,8 +71,10 @@ const GlobalNav = ({ selectedLink: selected, subset, tag }) => {
   const retrievedAvatar = useSelector((state) => state?.user?.profile?.avatar_url)
 
   const featureState = useSelector((state) => state.features)
-  const showHome = featureFlagActive({ flag: 'home.retention', featureState })
   const showLab = featureFlagActive({ flag: 'lab', featureState })
+  const showV1 = featureFlagActive({ flag: 'home.retention', featureState })
+  const showV2 = featureFlagActive({ flag: 'home.v2', featureState })
+  const showHome = showV1 || showV2
 
   const avatarSrc = enforceDefaultAvatar(retrievedAvatar)
   const accountName = useSelector((state) => state?.user?.first_name)