+++ src/pages/my-list/tags/[tag]/[filter].js
@@ -1,3 +1,14 @@
 import MyList from 'containers/my-list/my-list'
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
+import { LOCALE_COMMON } from 'common/constants'
+
+export async function getServerSideProps({ locale }) {
+  return {
+    props: {
+      authRequired: true,
+      ...(await serverSideTranslations(locale, [...LOCALE_COMMON]))
+    } // will be passed to the page component as props
+  }
+}
 
 export default MyList